fix MakeGPSPhoto function
